move paginated / sorted data to hook
