+++ src/hooks/use-screener.ts
@@ -1,8 +1,17 @@
 import { SearchQueryOption } from '@/components/stuff/ScreenerSearch.vue'
 import { Screener } from '@/interfaces/screener'
-import { NormalisedRow, UnknownObject, isValidInput, normaliseInput, omitFields, pickFields } from '../utils/data.utils'
+import {
+  NormalisedRow,
+  UnknownObject,
+  getPaginated,
+  isValidInput,
+  normaliseInput,
+  omitFields,
+  pickFields,
+} from '../utils/data.utils'
 import { computed, ref } from 'vue'
 import { search } from '../utils/search.utils'
+import { orderBy } from 'natural-orderby'
 
 const searchQuery = ref<string>('')
 const highlightQuery = ref<string>('')
@@ -54,6 +63,34 @@ export const useScreener = (options: ScreenerOptions = {}): Screener => {
     })
   })
 
+  const sortedData = computed((): NormalisedRow[] => {
+    const sortedRows = searchQuery.value ? searchedData.value : normalisedData.value
+
+    const sortIndex = sortedRows[0]?.findIndex((column) => column.key === sortField.value) ?? null
+
+    if (sortField.value && sortDirection.value) {
+      const nullRows = sortedRows.filter((row) => row?.[sortIndex] === null || row?.[sortIndex] === undefined)
+
+      const nonNullRows = sortedRows.filter((row) => row?.[sortIndex] !== null && row?.[sortIndex] !== undefined)
+
+      return [
+        ...orderBy(nonNullRows, [(row: NormalisedRow | null) => row?.[sortIndex]?.value], [sortDirection.value]),
+        ...nullRows,
+      ]
+    } else {
+      return sortedRows
+    }
+  })
+
+  const paginatedData = computed((): NormalisedRow[] => {
+    return getPaginated({
+      rows: sortedData.value,
+      page: currentPage.value - 1,
+      perPage: perPage.value,
+      withPlaceholders: true,
+    })
+  })
+
   return {
     searchQuery,
     highlightQuery,
@@ -69,6 +106,8 @@ export const useScreener = (options: ScreenerOptions = {}): Screener => {
     data,
     normalisedData,
     searchedData,
+    sortedData,
+    paginatedData,
     totalItems: computed(() => searchedData.value.length),
     actions: {
       search: (query: string, options?: SearchQueryOption[]) => {
